requirements.txt
tensorflow
keras
streamlit
opencv-python
pillow
numpy